Selection_sort.cpp
Algorithm for Selection Sort:

Start with the first element of the array as the current minimum.
Iterate through the array from the first element to the second-to-last element:
a. Find the index of the minimum element in the remaining unsorted portion.
b. If the minimum element is not the current element, swap them.
Repeat step 2 until the array is fully sorted.
Time Complexity:

Best Case: O(n^2)
Average Case: O(n^2)
Worst Case: O(n^2